bfloat16 for deepspeed

+++ train_oasis/utils.py
@@ -483,7 +483,7 @@ def log_video(
         observation_hat[:context_frames, :, i, :, [0, -1]] = c
         observation_gt[:, :, i, [0, -1], :] = c
         observation_gt[:, :, i, :, [0, -1]] = c
-    video = torch.cat([observation_hat, observation_gt], -1).detach().cpu().numpy()
+    video = torch.cat([observation_hat, observation_gt], -1).detach().to(torch.float).cpu().numpy()
     video = np.transpose(np.clip(video, a_min=0.0, a_max=1.0) * 255, (1, 0, 2, 3, 4)).astype(np.uint8)
     # video[..., 1:] = video[..., :1]  # remove framestack, only visualize current frame
     n_samples = len(video)
